utils: add check_process_exists method
Sometimes, some services need to be monitored only by checking if the
process is running well or not.
This patch add a new method for this and process.py to use this method.

The use case is for example Ceilometer Compute Agent where we currently
just need to validate the service is running as a process.